Development

+++ src/main/java/com/api/management/leave/leavemanagementapi/service/impl/EmployeeServiceImpl.java
@@ -2,13 +2,16 @@ package com.api.management.leave.leavemanagementapi.service.impl;
 
 import com.api.management.leave.leavemanagementapi.dto.EmployeeDto;
 import com.api.management.leave.leavemanagementapi.dto.EmployeeResponse;
+import com.api.management.leave.leavemanagementapi.dto.LeaveRequestDto;
+import com.api.management.leave.leavemanagementapi.dto.LeaveResponseDto;
 import com.api.management.leave.leavemanagementapi.entity.Employee;
 import com.api.management.leave.leavemanagementapi.exception.ResourceNotFoundException;
 import com.api.management.leave.leavemanagementapi.mapper.EmployeeMapper;
 import com.api.management.leave.leavemanagementapi.repository.EmployeeRepository;
 import com.api.management.leave.leavemanagementapi.service.EmployeeService;
 import com.api.management.leave.leavemanagementapi.utils.AppConstants;
-import lombok.RequiredArgsConstructor;
+import com.api.management.leave.leavemanagementapi.utils.LeaveTypes;
+import lombok.AllArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
@@ -17,11 +20,12 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.math.BigDecimal;
+import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Collectors;
 
 @Service
-@RequiredArgsConstructor
+@AllArgsConstructor
 public class EmployeeServiceImpl implements EmployeeService {
     private EmployeeRepository employeeRepository;
     private EmployeeMapper employeeMapper;
@@ -36,7 +40,9 @@ public class EmployeeServiceImpl implements EmployeeService {
         employee.setSickLeaveTotal(AppConstants.DEFAULT_SICK_LEAVE);
         employee.setLeaveWithoutPayTotal(AppConstants.DEFAULT_LEAVE_WITHOUT_PAY);
         Employee savedEmployee = employeeRepository.save(employee);
-        return employeeMapper.toDto(savedEmployee);
+        EmployeeDto employeeToDto = employeeMapper.toDto(savedEmployee);
+        employeeToDto.setMessage("Employee created successfully.");
+        return employeeToDto;
     }
 
     @Override
@@ -91,15 +97,18 @@ public class EmployeeServiceImpl implements EmployeeService {
 
     @Override
     public void deleteEmployee(Long id) {
-        employeeRepository.findById(id)
-                .orElseThrow(() -> new ResourceNotFoundException("Employee", "id", id));
+        boolean employee = employeeRepository.existsById(id);
+        if (!employee) {
+            throw new ResourceNotFoundException(AppConstants.EMPLOYEE, "Id", id);
+        }
+
         employeeRepository.deleteById(id);
     }
 
     @Override
-    public EmployeeDto excludeEmployeeForcedLeave(Long id, Boolean excluded) {
+    public EmployeeDto excludeEmployeeForcedLeave(Long id, boolean excluded) {
         Employee employee = employeeRepository.findById(id)
-                .orElseThrow(() -> new ResourceNotFoundException("Employee", "id", id));
+                .orElseThrow(() -> new ResourceNotFoundException(AppConstants.EMPLOYEE, "id", id));
         StringBuilder message = new StringBuilder();
         if (excluded) {
             if (employee.isExcluded()) {
@@ -147,11 +156,41 @@ public class EmployeeServiceImpl implements EmployeeService {
             response.setMessage("Employee leaves successfully reset.");
             return response;
         } else {
-            EmployeeResponse response = getAllEmployees(pageNo, pageSize, sortBy, sortDir);
+            EmployeeResponse response = this.getAllEmployees(pageNo, pageSize, sortBy, sortDir);
             response.setMessage("Employee leaves failed to reset.");
             return response;
         }
 
+    }
+
+    @Override
+    public LeaveResponseDto getEmployeeByOfficialEmailOrEmployeeNumber(String query) {
+        Employee employee = employeeRepository.findEmployeeByEmailOrEmployeeNumber(query)
+                .orElseThrow(() -> new ResourceNotFoundException(AppConstants.EMPLOYEE, "query", query));
+        EmployeeDto employeeDto =  employeeMapper.toDto(employee);
+        LeaveResponseDto leaveResponseDto = new LeaveResponseDto();
+        leaveResponseDto.setEmployeeDto(employeeDto);
+        List<String> leaveTypes = Arrays.stream(LeaveTypes.values()).map(leaveType -> leaveType.getLeave()).collect(Collectors.toList());
+        leaveResponseDto.setLeaveTypes(leaveTypes);
+        BigDecimal availableForcedLeaveToCancel = employee.getVacationLeaveTotal().subtract(employee.getRemainingForcedLeave());
+        availableForcedLeaveToCancel = availableForcedLeaveToCancel.signum() == -1
+                ? AppConstants.ZERO
+                : availableForcedLeaveToCancel;
+        leaveResponseDto.setAvailableForcedLeaveToCancel(availableForcedLeaveToCancel);
+        return leaveResponseDto;
+    }
+
+    @Override
+    public LeaveResponseDto leaveRequest(Long id, LeaveRequestDto leaveRequestDto) {
+        Employee employee = employeeRepository.findById(id)
+                .orElseThrow(() -> new ResourceNotFoundException(AppConstants.EMPLOYEE, "Id", id));
+
+        EmployeeDto employeeDto = employeeMapper.toDto(employee);
+        LeaveResponseDto leaveResponseDto = new LeaveResponseDto();
+        leaveResponseDto.setEmployeeDto(employeeDto);
+
+        // TODO: Process of Leave Request
+        return leaveResponseDto;
     }
 
     private EmployeeResponse setEmployeeResponse(Page<Employee> employees) {
@@ -170,11 +209,9 @@ public class EmployeeServiceImpl implements EmployeeService {
 
     private void computeForcedLeave(Employee employee) {
         BigDecimal remVacationLeave = employee.getVacationLeaveTotal();
-        if (remVacationLeave.compareTo(AppConstants.FIVE) == 1 || remVacationLeave.compareTo(AppConstants.FIVE) == 0) {
-            employee.setRemainingForcedLeave(AppConstants.FIVE);
-        } else {
-            employee.setRemainingForcedLeave(remVacationLeave);
-        }
-
+        employee.setRemainingForcedLeave(remVacationLeave.compareTo(AppConstants.FIVE) == 1
+                || remVacationLeave.compareTo(AppConstants.FIVE) == 0
+                ? AppConstants.FIVE
+                : remVacationLeave);
     }
 }
